refactor(vcs): share context-line slicing helper
Every VCS backend (git libgit2, git CLI, jj, hg, the no-VCS
FileBackend) and both forge backends (GitHub, GitLab) has their own
copy of the loop that slices file content into DiffLines for
gap-expansion context.

Extract it into vcs::slice_context_lines so it exists once.
