+++ src/forge/github/gh.rs
@@ -9,10 +9,11 @@ use crate::forge::traits::{
     PagedPullRequests, PullRequestCommit, PullRequestDetails, PullRequestListQuery,
     PullRequestListScope, PullRequestTarget,
 };
-use crate::model::{DiffLine, LineOrigin};
+use crate::model::DiffLine;
 use crate::process::{
     CommandOutputError, CommandOutputErrorKind, run_command_output, run_command_output_with_stdin,
 };
+use crate::vcs::slice_context_lines;
 
 use super::models::{GhPrCommit, GhPullRequestDetails, GhPullRequestSummary};
 use super::review_summaries::{
@@ -436,7 +437,7 @@ where
             self.fetch_file_via_api(&request)?
         };
 
-        Ok(slice_to_diff_lines(
+        Ok(slice_context_lines(
             &content,
             request.start_line,
             request.end_line,
@@ -580,25 +581,6 @@ where
         args.push(endpoint);
         self.run_gh(args, &request.repository.host)
     }
-}
-
-fn slice_to_diff_lines(content: &str, start_line: u32, end_line: u32) -> Vec<DiffLine> {
-    let lines: Vec<&str> = content.lines().collect();
-    let mut result = Vec::new();
-    for line_num in start_line..=end_line {
-        let idx = (line_num - 1) as usize;
-        if idx >= lines.len() {
-            break;
-        }
-        result.push(DiffLine {
-            origin: LineOrigin::Context,
-            content: lines[idx].to_string(),
-            old_lineno: Some(line_num),
-            new_lineno: Some(line_num),
-            highlighted_spans: None,
-        });
-    }
-    result
 }
 
 pub fn parse_pull_request_target(input: &str) -> Result<PullRequestTarget> {
